Update evl-emu.py
Add  option for supporting for legacy PC5401 module that has similar functionality as IT-100, but with less features.
Add option to keep the Stay arm as Stay arm. Some might want this function in stead of Zero entry.
Add option to disable the "Send Code" bug workaround. Some IT-100 modules do not have this bug and they actually work as in the documentation.